feat(timeslot): enforce alignment, overlap detection, and reactivate inactive slots
- Change price type from BigDecimal to Integer across entity, DTOs, and SQL
- Switch time format from HH:mm:ss to HH:mm via @JsonFormat
- Add 30-minute alignment and 30/60-minute duration validation
- Replace exact-match existence checks with range-based overlap detection
  for active time slots on create, update, and bulk create
- Reactivate inactive time slots with updated price on re-creation
- Apply pessimistic locking (SELECT FOR UPDATE) on write operations
- Handle HttpMessageNotReadableException for malformed JSON requests

BREAKING CHANGE: Time slot API now expects HH:mm time format and integer
prices instead of BigDecimal; overlapping active slots are rejected

+++ src/test/java/com/project/modules/booking/BookingServiceIntegrationTest.java
@@ -3,7 +3,6 @@ package com.project.modules.booking;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.math.BigDecimal;
 import java.time.*;
 import java.util.List;
 
@@ -105,6 +104,6 @@ class BookingServiceIntegrationTest {
 
     private TimeSlot createTimeSlot(Court court, LocalTime start, LocalTime end) {
         return timeSlots.save(TimeSlot.builder().court(court).startTime(start).endTime(end)
-                .price(BigDecimal.valueOf(50_000)).build());
+                .price(50_000).build());
     }
 }
